+++ src/data/sample.rs
@@ -2,8 +2,8 @@ use ndarray::Array1;
 
 #[derive(Debug, Clone)]
 pub struct Sample {
-    pub inputs: Array1<f64>,
-    pub expected: Array1<f64>,
+    pub inputs: Array1<f32>,
+    pub expected: Array1<f32>,
 }
 
 impl Sample {
